client/ezra/full-requirements.txt
anyio==4.9.0
arrow==1.3.0
binaryornot==0.4.4
blinker==1.9.0
briefcase==0.3.22
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
coverage==7.8.0
cryptography==44.0.2
dmgbuild==1.6.5
ds-store==1.3.1
Flask==3.1.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
mac-alias==2.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pipdeptree==2.26.0
platformdirs==4.3.7
pluggy==1.5.0
psutil==6.1.1
pycairo==1.27.0
pycparser==2.22
Pygments==2.19.1
PyGObject==3.50.0
pyproject_hooks==1.2.0
PySNARK @ git+https://github.com/meilof/pysnark@02eb721d7327ced1328783c820b25d5e39575dbc
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rubicon-objc==0.4.9
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
std-nslog==1.0.3
text-unidecode==1.3
toga==0.5.0
toga-cocoa==0.4.9
toga-core==0.4.9
toga-gtk==0.5.0
tomli_w==1.2.0
travertino==0.3.0
types-python-dateutil==2.9.0.20241206
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.46.1
zipp==3.21.0
